// apps/arabic-mandi/src/app/Core/admin/generate-admin-types.ts
export type Maybe<T> = T | null;
export type InputMaybe<T> = Maybe<T>;
export type Exact<T extends { [key: string]: unknown }> = { [K in keyof T]: T[K] };
export type MakeOptional<T, K extends keyof T> = Omit<T, K> & { [SubKey in K]?: Maybe<T[SubKey]> };
export type MakeMaybe<T, K extends keyof T> = Omit<T, K> & { [SubKey in K]: Maybe<T[SubKey]> };
/** All built-in and custom scalars, mapped to their actual values */
export type Scalars = {
  ID: string;
  String: string;
  Boolean: boolean;
  Int: number;
  Float: number;
};

export type BooleanFieldComparison = {
  is?: InputMaybe<Scalars['Boolean']>;
  isNot?: InputMaybe<Scalars['Boolean']>;
};
// ...
export type FoodCategoryInput = {
  category_image?: InputMaybe<Scalars['String']>;
  clientView?: InputMaybe<Scalars['Boolean']>;
  id?: InputMaybe<Scalars['Float']>;
  isActive?: InputMaybe<Scalars['Boolean']>;
  name?: InputMaybe<Scalars['String']>;
};
// ...
export type ItemEntityFilter = {
  and?: InputMaybe<Array<ItemEntityFilter>>;
  category?: InputMaybe<ItemEntityFilterFoodCategoryFilter>;
  createdby?: InputMaybe<ItemEntityFilterUserFilter>;
  foodsize?: InputMaybe<ItemEntityFilterFoodSizeFilter>;
  foodtype?: InputMaybe<ItemEntityFilterFoodTypeFilter>;
  id?: InputMaybe<NumberFieldComparison>;
  name?: InputMaybe<StringFieldComparison>;
  offer?: InputMaybe<StringFieldComparison>;
  or?: InputMaybe<Array<ItemEntityFilter>>;
  status?: InputMaybe<BooleanFieldComparison>;
  type?: InputMaybe<BooleanFieldComparison>;
  updatedby?: InputMaybe<ItemEntityFilterUserFilter>;
};

export type ItemEntityFilterFoodCategoryFilter = {
  and?: InputMaybe<Array<ItemEntityFilterFoodCategoryFilter>>;
  clientView?: InputMaybe<BooleanFieldComparison>;
  id?: InputMaybe<NumberFieldComparison>;
  isActive?: InputMaybe<BooleanFieldComparison>;
  name?: InputMaybe<StringFieldComparison>;
  or?: InputMaybe<Array<ItemEntityFilterFoodCategoryFilter>>;
};

export type ItemEntityFilterFoodSizeFilter = {
  and?: InputMaybe<Array<ItemEntityFilterFoodSizeFilter>>;
  id?: InputMaybe<NumberFieldComparison>;
  isActive?: InputMaybe<BooleanFieldComparison>;
  name?: InputMaybe<StringFieldComparison>;
  or?: InputMaybe<Array<ItemEntityFilterFoodSizeFilter>>;
};

export type ItemEntityFilterFoodTypeFilter = {
  and?: InputMaybe<Array<ItemEntityFilterFoodTypeFilter>>;
  id?: InputMaybe<NumberFieldComparison>;
  isActive?: InputMaybe<BooleanFieldComparison>;
  name?: InputMaybe<StringFieldComparison>;
  or?: InputMaybe<Array<ItemEntityFilterFoodTypeFilter>>;
};

export type ItemEntityFilterUserFilter = {
  and?: InputMaybe<Array<ItemEntityFilterUserFilter>>;
  id?: InputMaybe<NumberFieldComparison>;
  or?: InputMaybe<Array<ItemEntityFilterUserFilter>>;
  role?: InputMaybe<StringFieldComparison>;
  status?: InputMaybe<BooleanFieldComparison>;
};

export type ItemEntityMaxAggregate = {
  __typename?: 'ItemEntityMaxAggregate';
  id?: Maybe<Scalars['Float']>;
  name?: Maybe<Scalars['String']>;
  offer?: Maybe<Scalars['String']>;
};

export type ItemEntityMinAggregate = {
  __typename?: 'ItemEntityMinAggregate';
  id?: Maybe<Scalars['Float']>;
  name?: Maybe<Scalars['String']>;
  offer?: Maybe<Scalars['String']>;
};

export type ItemEntitySort = {
  direction: SortDirection;
  field: ItemEntitySortFields;
  nulls?: InputMaybe<SortNulls>;
};

export enum ItemEntitySortFields {
  Id = 'id',
  Name = 'name',
  Offer = 'offer',
  Status = 'status',
  Type = 'type'
}
// ...
export type NumberFieldComparison = {
  between?: InputMaybe<NumberFieldComparisonBetween>;
  eq?: InputMaybe<Scalars['Float']>;
  gt?: InputMaybe<Scalars['Float']>;
  gte?: InputMaybe<Scalars['Float']>;
  in?: InputMaybe<Array<Scalars['Float']>>;
  is?: InputMaybe<Scalars['Boolean']>;
  isNot?: InputMaybe<Scalars['Boolean']>;
  lt?: InputMaybe<Scalars['Float']>;
  lte?: InputMaybe<Scalars['Float']>;
  neq?: InputMaybe<Scalars['Float']>;
  notBetween?: InputMaybe<NumberFieldComparisonBetween>;
  notIn?: InputMaybe<Array<Scalars['Float']>>;
};
// ...
export type OfferItemInput = {
  category?: InputMaybe<FoodCategoryInput>;
  id?: InputMaybe<Scalars['Float']>;
  image_data?: InputMaybe<Scalars['String']>;
  name?: InputMaybe<Scalars['String']>;
  quantity?: InputMaybe<Scalars['Float']>;
  selected?: InputMaybe<Scalars['Boolean']>;
};
// ...
export enum SortDirection {
  Asc = 'ASC',
  Desc = 'DESC'
}

/** Sort Nulls Options */
export enum SortNulls {
  NullsFirst = 'NULLS_FIRST',
  NullsLast = 'NULLS_LAST'
}

export type StringFieldComparison = {
  eq?: InputMaybe<Scalars['String']>;
  gt?: InputMaybe<Scalars['String']>;
  gte?: InputMaybe<Scalars['String']>;
  iLike?: InputMaybe<Scalars['String']>;
  in?: InputMaybe<Array<Scalars['String']>>;
  is?: InputMaybe<Scalars['Boolean']>;
  isNot?: InputMaybe<Scalars['Boolean']>;
  like?: InputMaybe<Scalars['String']>;
  lt?: InputMaybe<Scalars['String']>;
  lte?: InputMaybe<Scalars['String']>;
  neq?: InputMaybe<Scalars['String']>;
  notILike?: InputMaybe<Scalars['String']>;
  notIn?: InputMaybe<Array<Scalars['String']>>;
  notLike?: InputMaybe<Scalars['String']>;
};
// ...
export type UpdateOfferInput = {
  discount?: InputMaybe<Scalars['Float']>;
  items?: InputMaybe<Array<OfferItemInput>>;
  name?: InputMaybe<Scalars['String']>;
  price?: InputMaybe<Scalars['Float']>;
  status?: InputMaybe<Scalars['Boolean']>;
  totalPrice?: InputMaybe<Scalars['Float']>;
  updatedby?: InputMaybe<UserInput>;
};
// ...
export type UpdateOneOfferInput = {
  /** The id of the record to update */
  id: Scalars['ID'];
  /** The update to apply. */
  update: UpdateOfferInput;
};
// ...
export type UserInput = {
  createdBy?: InputMaybe<Scalars['String']>;
  id?: InputMaybe<Scalars['Float']>;
  name?: InputMaybe<Scalars['String']>;
  profile_img?: InputMaybe<Scalars['String']>;
  role?: InputMaybe<Scalars['String']>;
  updatedBy?: InputMaybe<Scalars['String']>;
};
// ...
export type UpdateOneOfferMutationVariables = Exact<{
  input: UpdateOneOfferInput;
}>;
// ...
export type GetItemEntitiesQueryVariables = Exact<{
  filter: ItemEntityFilter;
  sorting?: InputMaybe<Array<ItemEntitySort> | ItemEntitySort>;
}>;